pyproject.toml
[project]
name = "mobile-rest-call"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    # "mysql-connector-python>=9.3.0",
    "aiomysql>=0.2.0",
    "daft>=0.5.7",
    "mysqlclient>=2.2.7",
    "numpy>=2.3.1",
    "pendulum>=3.1.0",
    "polars>=1.31.0",
    "psutil>=7.0.0",
    "pyarrow>=20.0.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
]
